Move functions and string_cache to top level
