+++ py-polars/polars/__init__.py
@@ -1,14 +1,18 @@
+# flake8: noqa
+
 # mypy needs these imported explicitly
 from polars.eager.frame import DataFrame, wrap_df  # noqa: F401
 from polars.eager.series import Series, wrap_s  # noqa: F401
 from polars.lazy.expr import Expr, wrap_expr  # noqa: F401
 from polars.lazy.frame import LazyFrame, wrap_ldf  # noqa: F401
 
-from . import datatypes, eager, io, lazy
+from . import datatypes, eager, functions, io, lazy, string_cache
 from .datatypes import *  # noqa: F401, F403
 from .eager import *  # noqa: F401, F403
+from .functions import *
 from .io import *  # noqa: F401, F403
 from .lazy import *  # noqa: F401, F403
+from .string_cache import *  # noqa: F401, F403
 
 # during docs building the binary code is not yet available
 try:
@@ -18,4 +22,11 @@ try:
 except ImportError:
     pass
 
-__all__ = datatypes.__all__ + eager.__all__ + io.__all__ + lazy.__all__
+__all__ = (
+    datatypes.__all__
+    + eager.__all__
+    + functions.__all__
+    + io.__all__
+    + lazy.__all__
+    + string_cache.__all__
+)
