+++ py-polars/polars/lazy/__init__.py
@@ -1,7 +1,8 @@
+# flake8: noqa
 from . import expr, frame, functions, whenthen
-from .expr import *  # noqa: F401, F403
-from .frame import *  # noqa: F401, F403
-from .functions import *  # noqa: F401, F403
-from .whenthen import *  # noqa: F401, F403
+from .expr import *
+from .frame import *
+from .functions import *
+from .whenthen import *
 
 __all__ = expr.__all__ + functions.__all__ + frame.__all__ + whenthen.__all__
